+++ src/main/java/github/renderbr/hytale/listeners/ServerStateListener.java
@@ -7,10 +7,12 @@ import com.hypixel.hytale.logger.backend.HytaleLoggerBackend;
 import com.hypixel.hytale.server.core.Message;
 import com.hypixel.hytale.server.core.event.events.BootEvent;
 import com.hypixel.hytale.server.core.event.events.ShutdownEvent;
+import com.hypixel.hytale.server.core.universe.Universe;
 import com.hypixel.hytale.server.core.universe.world.events.AllWorldsLoadedEvent;
 import github.renderbr.hytale.AverageDiscord;
 import github.renderbr.hytale.config.obj.ChannelOutputTypes;
 import github.renderbr.hytale.models.log.EventDrivenLogList;
+import github.renderbr.hytale.services.DiscordBotService;
 
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.logging.LogRecord;
@@ -26,6 +28,13 @@ public class ServerStateListener {
     }
 
     public static void onServerStart(BootEvent event) {
+
+        try {
+            AverageDiscord.instance = DiscordBotService.start();
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+
         AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.SERVER_STATE, Message.translation("server.bot.averagediscord.serverstarted").getAnsiMessage());
         AverageDiscord.instance.updateDiscordInformation();
     }
@@ -35,7 +44,17 @@ public class ServerStateListener {
     }
 
     public static void onLogReceived(LogRecord record) {
+        if (AverageDiscord.instance == null) {
+            return;
+        }
+
         String logMessage = record.getMessage();
+
+        // limit to 1500 characters to prevent crash
+        if (logMessage.length() > 1500) {
+            logMessage = logMessage.substring(0, 1500);
+        }
+
         AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.INTERNAL_LOG, logMessage);
     }
 }