+++ src/main/java/github/renderbr/hytale/registries/ListenerRegistry.java
@@ -1,9 +1,11 @@
 package github.renderbr.hytale.registries;
 
 import com.hypixel.hytale.event.EventRegistry;
+import github.renderbr.hytale.AverageDiscord;
 import github.renderbr.hytale.listeners.ChatListener;
 import github.renderbr.hytale.listeners.PlayerStateListener;
 import github.renderbr.hytale.listeners.ServerStateListener;
+import github.renderbr.hytale.services.DiscordBotService;
 
 public class ListenerRegistry {
     public static void registerListeners(EventRegistry eventRegistry){